fix: 修改医生、医院和文件实体的字段名称为下划线格式，并更新相关服务和控制器

+++ src/modules/doctor/doctor.service.ts
@@ -15,23 +15,23 @@ export class DoctorService {
   async addDoctor(userId: number, body: AddDoctorReqDto) {
     const { doctorName, six, hospitalId } = body;
     const newDoctor = new DoctorEntity();
-    newDoctor.doctorName = doctorName;
+    newDoctor.doctor_name = doctorName;
     newDoctor.six = six ?? '';
-    newDoctor.hospitalId = hospitalId;
-    newDoctor.userId = userId;
+    newDoctor.hospital_id = hospitalId;
+    newDoctor.user_id = userId;
     await this.hospitalRepository.save(newDoctor);
   }
 
   // 删除一个医院
   async deleteDoctor(userId: number, doctorId: number) {
-    await this.hospitalRepository.delete({ userId, id: doctorId });
+    await this.hospitalRepository.delete({ user_id: userId, id: doctorId });
     return await this.getExistDoctor(userId);
   }
 
   // 获取已有医院列表
   async getExistDoctor(userId: number) {
     const dbRes = await this.hospitalRepository.find({
-      where: { userId },
+      where: { user_id: userId },
     });
 
     return dbRes;
